Don't create backup of nonexistent files.
Backup of configuration files will fail if they do not exist.

+++ configure_nbextensions.py
@@ -29,7 +29,9 @@ def remove_old_config(configdata):
 def make_backup(filename):
     import shutil
     backup = filename + ".bak"
-    shutil.copy(filename,backup)
+    
+    if os.path.isfile(filename):
+        shutil.copy(filename, backup)
 
 
 def update_config(config_file):
